Refactor header dependencies
