+++ Source/OnixDevice.h
@@ -36,9 +36,6 @@
 
 #include "I2CRegisterContext.h"
 #include "NeuropixComponents.h"
-#include "UI/SettingsInterface.h"
-
-class SettingsInterface;
 
 using namespace std::chrono;
 
@@ -109,13 +106,10 @@ public:
 
 	int checkLinkState(oni_dev_idx_t port) const;
 
-	void setSettingsInterface(SettingsInterface* interface_);
-
 protected:
 
 	const oni_dev_idx_t deviceIdx;
 	const oni_ctx ctx;
-	SettingsInterface* settingsInterface;
 
 private:
